backend/requirements.txt
fastapi
pydantic
datetime
python-jose
passlib
uvicorn
python-multipart
reportlab
